# Optimized MemPostings.EnsureOrder()