While profiling Cortex ingesters at startup I've noticed `MemPostings.EnsureOrder()` scheduling overload is huge. For example, look at this profile:

![Screenshot 2021-11-05 at 09 16 54](https://user-images.githubusercontent.com/1701904/140479647-2f9dea5e-80ba-4285-b598-c37222d93936.png)

The CPI profile was taken over a 25s period and measured about 5.7 CPU cores utilization. During this timeframe 70% of CPU was spent in scheduling while `MemPostings.EnsureOrder()` was running.

The reason is that `EnsureOrder()` starts N workers and pushes single postings to each worker (via a channel). Sorting a single postings list is very fast but we typically have many of them to sort.

In this PR I'm proposing to use a batching strategy. I've added a benchmark to show the improvement:

```
name                                                 old time/op    new time/op    delta
MemPostings_ensureOrder/many_values_per_label-12        1.66s ± 3%     1.05s ± 2%  -36.73%  (p=0.001 n=3+3)
MemPostings_ensureOrder/few_values_per_label-12         1.86s ± 2%     1.14s ± 5%  -38.81%  (p=0.000 n=3+3)
MemPostings_ensureOrder/few_refs_per_label_value-12     697ms ± 3%      34ms ± 4%  -95.08%  (p=0.000 n=3+3)

name                                                 old alloc/op   new alloc/op   delta
MemPostings_ensureOrder/many_values_per_label-12       56.0MB ± 0%    24.6MB ± 0%  -56.06%  (p=0.000 n=3+3)
MemPostings_ensureOrder/few_values_per_label-12        56.0MB ± 0%    24.6MB ± 0%  -56.09%  (p=0.000 n=3+3)
MemPostings_ensureOrder/few_refs_per_label_value-12    56.0MB ± 0%    24.1MB ± 0%  -57.01%  (p=0.000 n=3+3)

name                                                 old allocs/op  new allocs/op  delta
MemPostings_ensureOrder/many_values_per_label-12        2.00M ± 0%     1.00M ± 0%  -49.95%  (p=0.000 n=3+3)
MemPostings_ensureOrder/few_values_per_label-12         2.00M ± 0%     1.00M ± 0%  -49.95%  (p=0.000 n=3+3)
MemPostings_ensureOrder/few_refs_per_label_value-12     2.00M ± 0%     1.00M ± 0%  -49.95%  (p=0.000 n=3+3)
```